👕 code refactor
